Feat: 가민 데이터 수집 추가
- 가민 데이터 수집 서비스 (백그라운드 + 스케줄러)
    - 일일 데이터 수집 및 DB 저장
    - 수집 시간 기준 설정
    - 분석용 데이터 적재
    - 배치 작업으로 실행
- 파티셔닝 매니저
    - 파티션 관리 작업 추가
- Celery 설정
    - RabbitMQ 연결
- 수동 작업 스크립트 설정

+++ backend/app/domain/daily_summary.py
@@ -91,8 +91,8 @@ class DailySummary(Data):
     # [층수]
     floors_ascended_in_meters: float
     floors_descended_in_meters: float
-    floors_ascended: int
-    floors_descended: int
+    floors_ascended: float
+    floors_descended: float
     user_floors_ascended_goal: int
 
     # [심박수]
